Replaced 'long *' by 'int *' or 'uint32_t *' for 64-bit compatibility.
git-svn-id: https://dosemu.svn.sourceforge.net/svnroot/dosemu/trunk@1723 7758b857-130e-0410-ab9c-d8735a528d18

+++ src/base/bios/setup.c
@@ -33,7 +33,7 @@ static void install_int_10_handler (void)
   
   if (config.vbios_seg == 0xe000 && config.vbios_post) {
     ptr = (u_char *)((BIOSSEG << 4) + ((long)bios_f000_int10ptr - (long)bios_f000));
-    *((long *)ptr) = 0xe0000003;
+    *((uint32_t *)ptr) = 0xe0000003;
     v_printf("VID: new int10 handler at %p\n",ptr);
   }
   else
