+++ src/emu-i386/simx86/fp87-x86.c
@@ -726,7 +726,7 @@ fpp002:
 			p[3] = p[4] = p[5] = p[6] = 0; q = (char *)(p+7);
 		    }
 		    else {
-			unsigned long *p = (unsigned long *)TheCPU.mem_ref;
+			uint32_t *p = (uint32_t *)TheCPU.mem_ref;
 			p[0] = TheCPU.fpuc; p[1] = TheCPU.fpus; p[2] = TheCPU.fptag;
 			/* IP,OP,opcode: n.i. */
 			p[3] = p[4] = p[5] = p[6] = 0; q = (char *)(p+7);
